Castep: refactor Cell file reader
Codacy keeps grumbling about how complex these methods are. Well,
there's a bit of redunancy in the treatment of line-reading that can
be factored out for clarity. However, there is a fundamental challenge
here that DFT input files tend to be complex with many branching
possibilities and it is not always going to be possible to make a
parser without a similar amount of logic...