Castep: CastepPhonon class can be instantiated from file
File reader implemented for CastepPhonon from .phonon file.
Header is read first to determine data sizes, line counts etc.
Then header is passed to main band parser for qpts, frequencies and
complex eigenvectors. (Not that we have a plan to use the
eigenvectors, but it's simpler to implement it now.)

Next step: construct a pymatgen phonon band structure from this object.